refactor: share normalization cache
